fix: fix how exclusions can be made
I discovered that if something is in the "include" array, you cannot remove it via an element in the `exclude` array unless it matches exactly.

As such, I've moved exclusion matching into the container itself.
A filter is run after gathering jobs, matching each against the exclusions to determine which to remove.
Any removed are logged for reference.

As such, we no longer emit an "exclude" array in the matrix.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>